+++ Jeffrey/Jeffrey/Scenes/Login/ViewModel/LoginViewModel.swift
@@ -1,5 +1,10 @@
 import Foundation
 import UIKit
+import Firebase
+import FirebaseAuth
+import FacebookLogin
+import FacebookCore
+import GoogleSignIn
 
 protocol LoginViewModelProtocol: AnyObject{
     func loginEmail()
@@ -7,7 +12,7 @@ protocol LoginViewModelProtocol: AnyObject{
     func openHome()
     func googleLogin()
     func appleLogin()
-    func facebookLogin()
+    func facebookLogin(viewController: UIViewController)
     var viewController: LoginViewEvents? {get set}
 }
 
@@ -37,14 +42,34 @@ class LoginViewModel: LoginViewModelProtocol{
     }
     
     func googleLogin() {
-        
+        GIDSignIn.sharedInstance()?.signIn()
     }
     
     func appleLogin() {
         //abrir login via apple
     }
     
-    func facebookLogin() {
-        //abir login via facebook
+    func facebookLogin(viewController: UIViewController) {
+        let manager = LoginManager()
+        manager.logIn(permissions: [.publicProfile, .email], viewController: viewController) { (result) in
+            switch result {
+            case .success(granted: let grantedPermissions, declined: let declinedPermission, token: let acessToken):
+                let credential = FacebookAuthProvider.credential(withAccessToken: acessToken.tokenString)
+                Auth.auth().signIn(with: credential) { (result, error) in
+                    guard  let homeViewControler = UIStoryboard(name: "HomeMain", bundle: nil).instantiateInitialViewController() as? UITabBarController else { return }
+                    DispatchQueue.main.async {
+                        
+                        UIViewController.replaceRootViewController(viewController: homeViewControler)
+                    }
+                }
+                print("acess Token = \(acessToken)")
+            case .cancelled:
+                print("Usuário cancelou o processo de Login")
+                break
+            case .failed(let error):
+                print("Login falhou! \(error.localizedDescription)")
+            }
+            manager.logOut()
+        }
     }
 }
